anpassungen der Benutzer & Buchungen = Controller und Services + herstellung Login Model und einbau in Benutzer

+++ src/main/java/ch/zli/m223/service/BenutzerService.java
@@ -1,14 +1,18 @@
 package ch.zli.m223.service;
+
 import java.util.List;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
+import javax.persistence.EntityExistsException;
 import javax.persistence.EntityManager;
+import javax.persistence.TransactionRequiredException;
 import javax.transaction.Transactional;
 
 import ch.zli.m223.exceptions.InvalidLoginException;
+import ch.zli.m223.exceptions.NullValueException;
 import ch.zli.m223.model.Benutzer;
-import ch.zli.m223.security.jwt.*;
+import ch.zli.m223.model.Login;
 
 @ApplicationScoped
 public class BenutzerService {
@@ -21,18 +25,34 @@ public class BenutzerService {
     }
 
     @Transactional
-    public Benutzer createUser(Benutzer benutzer) {
-        entityManager.persist(benutzer);
+    public Benutzer createBenutzer(Benutzer benutzer) throws EntityExistsException, Exception {
+        try {
+            entityManager.persist(benutzer);
+        } catch (EntityExistsException e) {
+            throw e;
+        } catch (Exception e) {
+            throw e;
+        }
+
         return benutzer;
     }
 
     @Transactional
-    public Benutzer updateUser(Benutzer benutzer) {
-        return entityManager.merge(benutzer);
+    public Benutzer updateBenutzer(Benutzer benutzer) throws IllegalArgumentException, TransactionRequiredException {
+        try {
+            return entityManager.merge(benutzer);
+        } catch (IllegalArgumentException e) {
+            throw e;
+        } catch (TransactionRequiredException e) {
+            throw e;
+        }
     }
 
-    public void deleteUser(Long id) {
+    public void deleteBenutzer(Long id) throws NullValueException {
         var entity = entityManager.find(Benutzer.class, id);
+        if (entity == null) {
+            throw new NullValueException("No user with id: " + id + " was found");
+        }
         entityManager.remove(entity);
     }
 
@@ -48,8 +68,14 @@ public class BenutzerService {
         return "THIS IS A TOKEN";
     }
     
-    public String register(Benutzer benutzer) {
-        createUser(benutzer);
+    public String register(Benutzer benutzer) throws EntityExistsException, Exception {
+        try {
+            createBenutzer(benutzer);
+        } catch (EntityExistsException e) {
+            throw e;
+        } catch (Exception e) {
+            throw e;
+        }
 
         //TODO: return real token
         return "This is a token";
